Reverting some changes
